enable to inject input file
