+++ onix/message.go
@@ -1,12 +1,7 @@
-package main
+package onix
 
 import (
-	"bytes"
 	"encoding/xml"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"log"
 )
 
 // Header is not documented yet.
@@ -145,29 +140,4 @@ type IngramContentOnix struct {
 	XMLName  xml.Name  `xml:"ONIXmessage"`
 	Header   Header    `xml:"header"`
 	Products []Product `xml:"product"`
-}
-
-func main() {
-	file, err := ioutil.ReadFile("/Users/kogaishinichi/k9books/onix/14362217.onix")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	var data IngramContentOnix
-	decoder := xml.NewDecoder(bytes.NewReader(file))
-	decoder.CharsetReader = func(label string, input io.Reader) (io.Reader, error) {
-		return input, nil
-	}
-
-	if err := decoder.Decode(&data); err != nil {
-		log.Fatal(err)
-	}
-	// fmt.Println("data=", data)
-	for _, d := range data.Products {
-		fmt.Println("d=", d.PublishingStatus)
-		// d.PublishingStatus
-		// for _, t := range d.OtherText {
-		// 	fmt.Println("t=", t.Text)
-		// }
-	}
 }
